+++ helper_js/pdf-link-scraper.js
@@ -1,9 +1,14 @@
-const pdfRegex = new RegExp("http.*\.pdf");
+const pdfRegex = new RegExp(/iframe\s*src\s*=\s*"(.*?\.pdf)/);
 
 function getPdfDownloadLink(htmlSource) {
   foundRegex = htmlSource.match(pdfRegex);
+  console.log(foundRegex);
   if (foundRegex) {
-    return foundRegex[0];
+    toReturn = foundRegex[1];
+    if (toReturn.startsWith("//")) { // quirk of sci-hub.st
+      toReturn = "https:" + toReturn;
+    }
+    return toReturn;
   }
 }